modif visuelles page d'accueil + re création page d'accueil en dehors de la page logo view
