+++ PageLogoView.swift
@@ -10,7 +10,7 @@ import UIKit
 
 struct PageLogoView: View {
     
-//    let imgswu : String="Logo-Appli"
+    //    let imgswu : String="Logo-Appli"
     @State private var hideFirstView = false
     
     let text = "Cyberhelp"
@@ -32,7 +32,7 @@ struct PageLogoView: View {
             if !hideFirstView {
                 Color("Neutre")
                     .ignoresSafeArea()
-               HStack {
+                HStack {
                     Image("LogoSeul")
                         .resizable()
                         .frame(width: 80, height: 80)
@@ -46,7 +46,7 @@ struct PageLogoView: View {
                     animateText()
                     DispatchQueue.main.asyncAfter(deadline: .now() + 2) {
                         withAnimation(.easeIn(duration: 1)) {
-                         
+                            
                             hideFirstView = true
                         }
                     }
@@ -55,128 +55,9 @@ struct PageLogoView: View {
                     
                 }.opacity(hideFirstView ? 0 : 1)
             } else {
-                        
-                AccueilView()
-                            
-                }
-            }
-        }
-    
-    
-    struct AccueilView: View {
-        
-        
-        @EnvironmentObject var ArticlesActuVM : ListArticlesActu
-        @EnvironmentObject var ArticlesEventsVM : ListArticlesEvents
-        @EnvironmentObject var ArticlesTemoignagesVM : ListArticlesTemoignages
-        
-        @State private var selectedTopic: TopicOption = .actu
-        @State var isShown : Bool = false
-        @State var addTemoignage : Bool = false
-        
-        init() {
-            UISegmentedControl.appearance().selectedSegmentTintColor = UIColor.white
-            UISegmentedControl.appearance().setTitleTextAttributes([.foregroundColor: UIColor(named: "Primaire")], for: .selected)
-            UISegmentedControl.appearance()
-                .backgroundColor = UIColor(named: "Primaire")
-            UISegmentedControl.appearance().setTitleTextAttributes([.foregroundColor: UIColor.white], for: .normal)
-            
-        }
-        
-        var body: some View {
-            NavigationView{
-                ZStack{
-                    Color("Neutre")
-                        .ignoresSafeArea()
-                    
-                    ScrollView {
-                        VStack{
-                            
-                            Picker((""), selection: $selectedTopic) {
-                                ForEach(TopicOption.allCases, id: \.self) { topic in
-                                    Text(topic.rawValue)
-                                }
-                            }
-                        }
-                        .pickerStyle(.segmented)
-                        .padding(.leading)
-                        .padding(.trailing)
-                        .foregroundColor(Color("DeepBlue"))
-                        
-                        VStack (alignment: .leading){
-                            
-                            ForEach (getList()) {
-                                displayActu in
-                                if displayActu.id == getList().first!.id{
-                                    NavigationLink {
-                                        ArticlesDetailsView(article: displayActu)
-                                    } label: {
-                                        ZStack (alignment: .bottom){
-                                            Image(displayActu.image)
-                                                .resizable()
-                                                .frame(maxWidth: .infinity)
-                                                .frame(height: 200)
-                                            Text(displayActu.titre)
-                                                .padding(8)
-                                                .foregroundColor(Color("DeepBlue"))
-                                                .font(.system(size: 16))
-                                                .fontWeight(.bold)
-                                                .frame(maxWidth: .infinity)
-                                                .padding(.horizontal)
-                                                .background(Color.white)
-                                                .opacity(0.9)
-                                        }
-                                        .cornerRadius(24)
-                                        .padding(.bottom, 24)
-                                    }
-                                }
-                                else {
-                                    NavigationLink {
-                                        ArticlesDetailsView(article: displayActu)
-                                    } label: {
-                                        HStack {
-                                            Image(displayActu.image)
-                                                .resizable()
-                                                .scaledToFill()
-                                                .frame(width: 100, height: 100)
-                                                .clipped()
-                                            
-                                            Text(displayActu.titre)
-                                                .foregroundColor(Color("DeepBlue"))
-                                                .multilineTextAlignment(.leading) .font(.system(size: 16))
-                                                .fontWeight(.bold)
-                                                .frame(width: 260)
-                                        }
-                                        .background(Color.white)
-                                        .cornerRadius(10)
-                                        .frame(width: 370)
-                                    }
-                                }
-                            }
-                        }
-                        .padding()
-                    }
-                    .navigationTitle("Bonjour Éléonore")
-                    .navigationBarTitleDisplayMode(.inline)
-                    
-                    if selectedTopic == .temoignages {
-                        AddButtonTemoignageView(addTemoignage: $addTemoignage)
-                    }
-                }
-            }
-            .sheet(isPresented: $addTemoignage) {
-                       ModaleTemoignages(isModaleShown: $addTemoignage)
-            }
-        }
-        
-        func getList() -> [Article] {
-            switch selectedTopic {
-            case .actu:
-                return ArticlesActuVM.listActu
-            case .event:
-                return ArticlesEventsVM.listEvents
-            case .temoignages:
-                return ArticlesTemoignagesVM.listTemoignages
+                
+                TabviewU()
+                
             }
         }
     }
