+++ View/Profil/ProfilPView.swift
@@ -144,8 +144,6 @@ struct ProfilPView: View {
                 
             }
         }
-        
-    }
 
 
 struct ProfilPView_Previews: PreviewProvider {
