+++ View/TabView/TabviewU.swift
@@ -15,7 +15,7 @@ struct TabviewU: View {
     @State private var selectedIndex = 0
     var body: some View {
         TabView(selection: $selectedIndex) {
-            PageLogoView()
+            AccueilView()
                 .tabItem {
                     if selectedIndex == 0 {
                         Image("homeFilled")
